% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\docType{data}
\name{degree}
\alias{degree}
\title{Degree data sample}
\format{
A \code{data.table} with approximately 48,000 observations and
5 variables occupying 10.2 MB of memory.
\describe{
  \item{mcid}{Character, anonymized student identifier}

  \item{institution}{Character, anonymized institution name, e.g.,
      Institution A, Institution B, etc.}

  \item{term}{Character, academic year and term, format YYYYT, in which a
       student completes their program}

  \item{cip6}{Character, 6-digit CIP code of program in which a student
       earns a degree}

  \item{degree}{Character, type of degree awarded, e.g., Bachelor's Degree,
       Bachelor of Arts, Bachelor of Science, etc.}
}
}
\source{
\itemize{
  \item{Data provided by the MIDFIELD project:
        \url{https://engineering.purdue.edu/MIDFIELD}}
  \item{CIP codes from NCES IPEDS:
        \url{https://nces.ed.gov/ipeds/cipcode/}}
}
}
\usage{
data(degree)
}
\description{
Table of degree-related attributes of approximately 48,000 undergraduates
upon completing their programs. Each observation is a unique student
keyed by student ID.
}
\details{
The data in midfielddata are a proportionate stratified random sample of 12
institutions in the MIDFIELD research database. However, these are practice
data, not research data---suitable for learning about student-record
analysis, but not for drawing inferences about student performance.

Students in the \code{degree} table also have records in the
\link[midfielddata]{student}, \link[midfielddata]{course}, and
\link[midfielddata]{term} tables.

Terms are encoded YYYYT, where YYYY is the year of the Fall term that starts
the academic year and T = 1 (Fall), 2 (Winter, quarter systems only), 3
(Spring), and 4 (Summer) or 5 and 6 (Summer sessions 1 and 2). For example,
for the 1995--96 academic year, Fall '95 is encoded 19951, Spring '96 is
encoded 19953, and the first summer '96 term is encoded 19954. The MIDFIELD
research database also includes terms of one month duration, encoded with
letters A, B, C, etc., though none are included in this sample.

For program codes, midfielddata uses the 2010 version of the Classification
of Instructional Programs (CIP) . See midfieldr
\code{\link[midfieldr:cip]{cip}} for more information.
}
\keyword{datasets}
