% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\docType{data}
\name{course}
\alias{course}
\title{Course data sample}
\format{
A \code{data.frame} and \code{data.table} with 12 variables and approximately
3.3M observations of 97,555 unique students occupying 325 MB of memory:

\describe{

\item{mcid}{Character, anonymized student identifier, e.g.,
\code{"MCID3111142225"}.}

\item{institution}{Character, anonymized institution name, e.g.,
\code{"Institution B"}.}

\item{term_course}{Character, academic year and term, format \code{"YYYYT"}.}

\item{course}{Character, course name, e.g., \code{"Astrophysics III"},
\code{"Calculus For Social Science And Business"}, \code{"Corp Financial Rprtng 1"},
\code{"Environmental Sanitation II"}, \code{"Fitness and Wellness"},
\code{"Introductory Astronomy 2"}, \code{"Our Changing Environment"}, etc.}

\item{abbrev}{Character, course alpha identifier, e.g. \code{"AA"}, \code{"MATH"},
\code{"ACCT"}, \code{"EH"}, \code{"HES"}, \code{"ASTR"}, etc.}
\item{number}{Character, course numeric identifier, e.g. \code{"1104"}, \code{"1209"},
\code{"228"}, \code{"4047"}, etc.}

\item{section}{Character, course section identifier, from one to four
characters, e.g., \code{"1"}, \code{"2"}, \code{"01"}, \code{"14"}, \code{"001"}, \code{"040"}, \code{"785"},
\code{"H02"}, \code{"R01"}, \code{"300E"}, \code{"888R"}, etc.}

\item{type}{Character, predominant delivery method for this section, e.g.,
\code{"Blended"}, \code{"Distance Education"}, \code{"Face-to-Face"}, \code{"Online"}, etc.}

\item{faculty_rank}{Character, academic rank of the person teaching the
course, e.g., \code{"Assistant Professor"}, \code{"Associate Professor"},
\code{"Graduate Assistant"}, \code{"Visiting Faculty"}, etc.}

\item{hours_course}{Numeric, number of credit-hours for successful course
completion.}

\item{grade}{Character, course grade, e.g., \code{"A+"}, \code{"A"}, \code{"A-"}, \code{"B+"},
\code{"I"}, \code{"NG"}, etc.}

\item{discipline_midfield}{Character, a variable for grouping courses by
academic discipline assigned by the MIDFIELD data curator, e.g.,
\code{"Anthropology"}, \code{"Business"}, \code{"Computer Science"},
\code{"Engineering"}, \code{"Language and Literature"}, \code{"Mathematics"},
\code{"Visual and Performing Arts"}, etc.}

}
}
\source{
Data provided by the MIDFIELD project: \url{https://midfield.online/}
}
\usage{
data(course, package = "midfielddata")
}
\description{
Course information for approximately 98,000 undergraduates. One record per
student per course per term, keyed by student ID.
}
\details{
Terms are encoded \code{YYYYT}, where \code{YYYY} is the year at the start of the
academic year and \code{T} encodes the semester or quarter within an academic year
as Fall (\code{1}), Winter (\code{2}), Spring (\code{3}), and Summer (\code{4}, \code{5}, and \code{6}).
For example, for academic year 1995--96, Fall 95--96 is encoded \code{19951},
Spring 95--96 is encoded \code{19953}, and the first Summer 95-96 term is encoded
\code{19954}. The source database includes special month-long sessions encoded
with letters \code{A}, \code{B}, \code{C}, etc., though none are included in this sample.

The data in 'midfielddata' are practice data, suitable for learning to work
with Student Unit Records (SURs) generally. Unlike the MIDFIELD source data,
the data tables in 'midfielddata' are not research data; they are not
suitable for drawing inferences about program attributes or student
experiences.
}
\keyword{datasets}
