% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\docType{data}
\name{degree}
\alias{degree}
\title{Degree data sample}
\format{
A \code{data.frame} and \code{data.table} with 5 variables and 49,543
observations of unique students occupying 6 MB of memory:

\describe{

\item{mcid}{Character, anonymized student identifier, e.g.,
\code{"MCID3111142225"}.}

\item{institution}{Character, anonymized institution name, e.g.,
\code{"Institution B"}.}

\item{term_degree}{Character, academic year and term, format \code{"YYYYT"}, in
which a student completes their program.}

\item{cip6}{Character, 6-digit CIP code of program in which a student earns
a degree, e.g., \code{"090101"}, \code{"141201"}, \code{"260901"}, \code{"420101"}, etc.}

\item{degree}{Character, type of degree awarded, e.g.,
\code{"Bachelor of Arts in Geography"}, \code{"Bachelor of Science in Finance"},
etc.}

}
}
\source{
\itemize{
\item Data provided by the MIDFIELD project: \url{https://midfield.online/}
\item CIP codes adapted from NCES IPEDS: \url{https://nces.ed.gov/ipeds/cipcode/}
}
}
\usage{
data(degree, package = "midfielddata")
}
\description{
Degree information for approximately 50,000 undergraduates earning degrees.
One record per student per degree, keyed by student ID. A subset of the
98,000 undergraduates in the practice data.
}
\details{
Terms are encoded \code{YYYYT}, where \code{YYYY} is the year at the start of the
academic year and \code{T} encodes the semester or quarter within an academic year
as Fall (\code{1}), Winter (\code{2}), Spring (\code{3}), and Summer (\code{4}, \code{5}, and \code{6}).
For example, for academic year 1995--96, Fall 95--96 is encoded \code{19951},
Spring 95--96 is encoded \code{19953}, and the first Summer 95-96 term is encoded
\code{19954}. The source database includes special month-long sessions encoded
with letters \code{A}, \code{B}, \code{C}, etc., though none are included in this sample.

For program codes, 'midfielddata' uses the 2010 version of the Classification
of Instructional Programs (CIP). If 'midfieldr' is installed, type
\code{help(cip, package = "midfieldr")} for details.

The data in 'midfielddata' are practice data, suitable for learning to work
with Student Unit Records (SURs) generally. Unlike the MIDFIELD source data,
the data tables in 'midfielddata' are not research data; they are not
suitable for drawing inferences about program attributes or student
experiences.
}
\keyword{datasets}
