% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfielddata-package.R
\docType{package}
\name{midfielddata-package}
\alias{midfielddata-package}
\title{MIDFIELD Data Sample}
\source{
Data provided by MIDFIELD: \url{https://midfield.online/}.
}
\description{
Provides a data sample for practice working with individual undergraduate
student unit records (registrar's data) in 'R'. Data are organized in four
data tables --- \code{student}, \code{course}, \code{term}, and \code{degree} --- providing
anonymized records for approximately 98,000 students at three US institutions
from 1988 through 2018, keyed by student ID. These data are a proportionate
stratified sample of the MIDFIELD database. However, unlike MIDFIELD data,
the data in 'midfielddata' are not suitable for drawing inferences about
program attributes or student experiences --- these are practice data, not
research data.
}
\details{
\href{https://midfield.online/}{MIDFIELD} is a database containing (as of October,
2022) individual Student Unit Records (SURs) for 1.7M undergraduate students
at 19 US institutions from 1987 through 2018. Access to the MIDFIELD research
database is currently limited to MIDFIELD partner institutions, but a sample
of the data is available in the 'midfielddata' package.

\href{https://midfieldr.github.io/midfieldr/}{'midfieldr'} is a companion 'R'
package that provides tools and detailed procedures for working with MIDFIELD
research data and with 'midfielddata' practice data.

This work is supported by a grant from the US National Science Foundation
(EEC 1545667).
}
\concept{package}
