% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfielddata-package.R
\docType{package}
\name{midfielddata-package}
\alias{midfielddata-package}
\title{Sample of MIDFIELD Student Unit Record Data}
\source{
Data provided by the MIDFIELD project: \url{https://midfield.online/}
}
\description{
An 'R' data package for distributing practice Student Unit Record (SUR) data
and documentation. The source is the MIDFIELD database, which contains (as of
October, 2022) individual SUR data for 1.7M undergraduate students at 19 US
institutions from 1987 through 2018. 'midfielddata' provides a proportionate
stratified sample of MIDFIELD data: anonymized SURs for approximately 98,000
students at three US institutions from 1988 through 2018. Data are organized
in four tables: \code{student}, \code{course}, \code{term}, and \code{degree}.
}
\details{
The data in 'midfielddata' are practice data, suitable for learning to work
with SURs generally. Unlike the MIDFIELD source data, the data tables in
'midfielddata' are not research data; they are not suitable for drawing
inferences about program attributes or student experiences.

The \href{https://midfieldr.github.io/midfieldr/}{'midfieldr'} package
provides tools for working with 'midfielddata' practice data as well as the
MIDFIELD research data.
}
