% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfielddata-package.R
\docType{package}
\name{midfielddata-package}
\alias{midfielddata-package}
\title{MIDFIELD Data Sample}
\source{
2022 \href{https://midfield.online/}{MIDFIELD} database
}
\description{
This data package contains anonymized student-level records for 98,000
undergraduates from three US institutions from 1988 through 2018 organized in
four data tables keyed by student ID. Data at the "student-level" refers to
information collected by undergraduate institutions about individual students
including demographics, programs, academic standing, courses, grades, and
degrees. These data are a proportionate stratified sample of the MIDFIELD
database, but are not suitable for drawing inferences about program
attributes or student experiences---\code{midfielddata} provides practice data,
not research data.
}
\seealso{
Package \href{https://midfieldr.github.io/midfieldr/}{\code{midfieldr}}
for tools and methods for working with MIDFIELD data in \code{R}.
}
\concept{package}
