% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfielddata-package.R
\docType{package}
\name{midfielddata-package}
\alias{midfielddata-package}
\title{Sample of MIDFIELD Student Unit Record Data}
\source{
Data provided by the MIDFIELD project:
\url{https://midfield.online/}
}
\description{
Provides a sample of MIDFIELD data for practice working with
longitudinal, de-identified, individual student unit records.
}
\details{
MIDFIELD contains individual Student Unit Record (SUR) data for
1.7M students at 21 US institutions (as of June 2022). MIDFIELD is large
enough to permit grouping and summarizing by multiple characteristics,
enabling researchers to examine student characteristics (race/ethnicity,
sex, prior achievement) and curricular pathways (including coursework and
program) by institution and over time.

midfieldata provides a sample of these data (for practice) with
longitudinal SURs for 98,000 undergraduates at 12 institutions
from 1987–2016 organized in four data tables:
\itemize{
 \item {\href{https://midfieldr.github.io/midfielddata/reference/student.html}{student}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/course.html}{course}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/term.html}{term}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/degree.html}{degree}}
}

Tools to work with these data are provided in the R package
\href{https://midfieldr.github.io/midfieldr/}{midfieldr}.
}
