% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldterms.R
\docType{data}
\name{midfieldterms}
\alias{midfieldterms}
\title{Term data for 98,000 undergraduates}
\format{
\code{data.table} with 727,369 observations and 13 variables,
occupying 82.1 MB of memory. Each observation is one term for one student.
The variables are:
\describe{
  \item{id}{character, anonymized student identifier, MID followed by an 
      8-digit integer string, composite key to these data with \code{term}}
  \item{institution}{character, anonymized institution name, e.g., 
      Institution A, Institution B, etc.}
  \item{cip6}{character, 6-digit CIP code of program in which a student is 
      enrolled in a term}
  \item{term}{numeric, academic year and term, format YYYYT, composite key 
      to these data with \code{id}}
  \item{level}{character, 01 Freshman, 02 Sophomore, etc.}
  \item{standing}{character, academic standing, e.g., 
      Good Standing, Suspended, First Time in College, etc.}
  \item{coop}{character, coop Y or N}
  \item{hours_term_attempt}{numeric, credit hours attempted in the term}
  \item{hours_term}{numeric, credit hours earned in the term}
  \item{gpa_term}{numeric, term grade point average}
  \item{hours_cumul_attempt}{numeric, cumulative credit hours attempted}
  \item{hours_cumul}{numeric, cumulative credit hours earned}
  \item{gpa_cumul}{numeric, cumulative grade point average}
}
}
\source{
Data provided by the MIDFIELD project:
\url{https://engineering.purdue.edu/MIDFIELD}
}
\usage{
midfieldterms
}
\description{
Data frame of term attributes of 95,995 undergraduates with 727,369
observations and 13 academic term variables keyed by student ID and term.
Each observation is one term for one student. A stratified sample
of the MIDFIELD database. The term data has 1645 fewer student IDs than 
the other data sets in midfielddata. The cause will be corrected or 
explained as soon as possible.
}
\details{
Classification of Instructional Programs (CIP) codes are from the 2010 CIP. 
See midfieldr \code{\link[midfieldr:cip]{cip}} for more information. 

Terms are encoded YYYYT, where YYYY is the year of the Fall term that starts 
the academic year and T = 1 (Fall), 2 (Winter, quarter systems only), 3 
(Spring), and 4 (Summer) or 5 and 6 (Summer sessions 1 and 2). All terms of 
an academic year are encoded with the same YYYY value. For example, the 
terms of the academic year 2009-10 for an institution on a semester system 
are encoded 20091 (Fall) and 20093 (Spring) with summer terms encoded 20094 
or 20095 and 20096 depending on the number of sessions.
}
\examples{
\dontrun{
library(data.table)
midfieldterms
}

}
\keyword{datasets}
