% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\docType{data}
\name{student}
\alias{student}
\title{Student data sample}
\format{
A \code{data.frame} and \code{data.table} with 13 variables and 97,555
observations of unique students occupying 18 MB of memory:

\describe{

\item{mcid}{Character, anonymized student identifier, e.g.,
\code{"MCID3111142225"}.}

\item{institution}{Character, anonymized institution name, e.g.,
\code{"Institution B"}.}

\item{transfer}{Character, transfer status, possible values are
\code{"First-Time in College"}, \code{"First-Time Transfer"}.}

\item{hours_transfer}{Numeric, number of credit hours transferred (or
\code{NA}).}

\item{race}{Character, race/ethnicity, e.g., \code{"Asian"}, \code{"Black"},
\code{"International"}, \code{"Latine"}, etc.}

\item{sex}{Character, sex, possible values are \code{"Female"}, \code{"Male"},
\code{"Unknown"}.}

\item{age_desc}{Character, age group, possible values are \code{"25 and Older"},
\code{"Under 25"}.}

\item{us_citizen}{Character, US citizenship, possible values are \code{"No"},
\code{"Yes"}.}

\item{home_zip}{Character, home ZIP code (or \code{NA}), e.g., \code{"02056"}, \code{"20170"},
\code{"51301"}, \code{"80129"}, etc.}

\item{high_school}{Character, code for the last high school attended
before admission (or \code{NA}), e.g., \code{"060075"}, \code{"210512"}, \code{"431800"},
\code{"502195"}, etc.}

\item{sat_math}{Numeric, SAT mathematics test score (or \code{NA}).}

\item{sat_verbal}{Numeric, SAT reading test score (or \code{NA}).}

\item{act_comp}{Numeric, ACT composite test score (or \code{NA}).}

}
}
\source{
Data provided by the MIDFIELD project: \url{https://midfield.online/}
}
\usage{
data(student, package = "midfielddata")
}
\description{
Demographic information at matriculation for approximately 98,000
degree-seeking undergraduate students. One record per student, keyed by
student ID.
}
\details{
The data in 'midfielddata' are practice data, suitable for learning to work
with Student Unit Records (SURs) generally. Unlike the MIDFIELD source data,
the data tables in 'midfielddata' are not research data; they are not
suitable for drawing inferences about program attributes or student
experiences.
}
\keyword{datasets}
