% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\docType{data}
\name{term}
\alias{term}
\title{Term data sample}
\format{
A \code{data.frame} and \code{data.table} with 13 variables and approximately
640,000 observations of 97,555 unique students occupying 73 MB of memory:

\describe{

\item{mcid}{Character, anonymized student identifier, e.g.,
\code{"MCID3111142225"}.}

\item{institution}{Character, anonymized institution name, e.g.,
\code{"Institution B"}.}

\item{term}{Character, academic year and term, format \code{"YYYYT"}.}

\item{cip6}{Character, 6-digit CIP code of program in which a student is
enrolled in a term, e.g., \code{"090101"}, \code{"141201"}, \code{"260901"}, \code{"420101"},
etc.}

\item{level}{Character, academic level determined by cumulative number of
credit ours earned, e.g., \code{"01 First-year"}, \code{"02 Second-year"}, etc.}

\item{standing}{Character, academic standing, e.g., \code{"Good Standing"},
\code{"Academic Warning"}, etc.}

\item{coop}{Character, cooperative education term, possible values are
\code{"Yes"}, \code{"No."}}

\item{hours_term}{Numeric, credit hours earned in the term.}

\item{hours_term_attempt}{Numeric, credit hours attempted in the term.}

\item{hours_cumul}{Numeric, cumulative credit hours earned.}

\item{hours_cumul_attempt}{Numeric, cumulative credit hours attempted.}

\item{gpa_term}{Numeric, term grade point average.}

\item{gpa_cumul}{Numeric, cumulative grade point average.}

}
}
\source{
\itemize{
\item Data provided by the MIDFIELD project: \url{https://midfield.online/}
\item CIP codes adapted from NCES IPEDS: \url{https://nces.ed.gov/ipeds/cipcode/}
}
}
\usage{
data(term, package = "midfielddata")
}
\description{
Term information for approximately 98,000 undergraduates. One record per
student per enrolled term, keyed by student ID.
}
\details{
Terms are encoded \code{YYYYT}, where \code{YYYY} is the year at the start of the
academic year and \code{T} encodes the semester or quarter within an academic year
as Fall (\code{1}), Winter (\code{2}), Spring (\code{3}), and Summer (\code{4}, \code{5}, and \code{6}).
For example, for academic year 1995--96, Fall 95--96 is encoded \code{19951},
Spring 95--96 is encoded \code{19953}, and the first Summer 95-96 term is encoded
\code{19954}. The source database includes special month-long sessions encoded
with letters \code{A}, \code{B}, \code{C}, etc., though none are included in this sample.

For program codes, 'midfielddata' uses the 2010 version of the Classification
of Instructional Programs (CIP). If  'midfieldr' is installed, type
\code{help(cip, package = "midfieldr")} for details.

The data in 'midfielddata' are practice data, suitable for learning to work
with Student Unit Records (SURs) generally. Unlike the MIDFIELD research
database, the data tables in 'midfielddata' are not research data; they are
not suitable for drawing inferences about program attributes or student
experiences.
}
\keyword{datasets}
