% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\docType{data}
\name{term}
\alias{term}
\title{Student-level term data}
\format{
A \code{data.frame} and \code{data.table} with 13 variables and approximately
640,000 observations of 97,555 unique students occupying 73 MB of memory:

\describe{
\item{\code{mcid}}{Character, anonymized student identifier, e.g., \code{MCID3111142225}.}
\item{\code{institution}}{Character, de-identified institution name, e.g., Institution
A, Institution B, etc.}
\item{\code{term}}{Character, academic year and term, format YYYYT.}
\item{\code{cip6}}{Character, 6-digit CIP code of program in which a student is
enrolled in a term, e.g., \code{090101}, \code{141201}, \code{260901}, \code{420101},
etc.}
\item{\code{level}}{Character, academic level determined by cumulative number of
credit ours earned, e.g., \verb{01 First-year}, \verb{02 Second-year}, etc.}
\item{\code{standing}}{Character, academic standing, e.g., \verb{Good Standing},
\verb{Academic Warning}, etc.}
\item{\code{coop}}{Character, cooperative education term, possible values are
\code{Yes}, \code{No}.}
\item{\code{hours_term}}{Numeric, credit hours earned in the term.}
\item{\code{hours_term_attempt}}{Numeric, credit hours attempted in the term.}
\item{\code{hours_cumul}}{Numeric, cumulative credit hours earned.}
\item{\code{hours_cumul_attempt}}{Numeric, cumulative credit hours attempted.}
\item{\code{gpa_term}}{Numeric, term grade point average.}
\item{\code{gpa_cumul}}{Numeric, cumulative grade point average.}
}
}
\source{
2022 \href{https://midfield.online/}{MIDFIELD} database
}
\usage{
data(term)
}
\description{
Student-level term information for approximately 98,000 undergraduates, keyed
by student ID. Data at the "student-level" refers to information collected by
undergraduate institutions about individual students, for example, program
code, academic standing, and grade point average by term.
}
\details{
Term data are structured in block-record form, that is, records associated
with a particular ID can span multiple rows---one record per student per
term.

Terms are encoded \code{YYYYT}, where \code{YYYY} is the year at the start of the
academic year and \code{T} encodes the semester or quarter within an academic year
as Fall (\code{1}), Winter (\code{2}), Spring (\code{3}), and Summer (\code{4}, \code{5}, and \code{6}).
For example, for academic year 1995--96, Fall 95--96 is encoded \code{19951},
Spring 95--96 is encoded \code{19953}, and the first Summer 95-96 term is encoded
\code{19954}. The source database includes special month-long sessions encoded
with letters \code{A}, \code{B}, \code{C}, etc., though none are included in this sample.

For program codes, \code{midfielddata} uses the 2010 version of the Classification
of Instructional Programs (CIP). If the \code{midfieldr} package is installed and
loaded, type \code{?cip} for details.

The data in \code{midfielddata} are a proportionate stratified sample of the
MIDFIELD database, but are not suitable for drawing inferences about program
attributes or student experiences---\code{midfielddata} provides practice data,
not research data.
}
\examples{
\dontrun{

# Load data
data(term)

# Select specific rows and columns
rows_we_want <- term$mcid == "MCID3112192438"
cols_we_want <- c("mcid", "term", "cip6", "level", "standing", "gpa_cumul")

# View observations for this ID 
term[rows_we_want, cols_we_want] 

}
  
}
\seealso{
Package \href{https://midfieldr.github.io/midfieldr/}{\code{midfieldr}} for
tools and methods for working with MIDFIELD data in \code{R}.

Other datasets: 
\code{\link{course}},
\code{\link{degree}},
\code{\link{student}}
}
\concept{datasets}
\keyword{datasets}
